backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
yt-dlp==2024.4.9
openai-whisper>=20231117
openai==1.30.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
